Address PR #11 review feedback and stabilize cache tests
